refactor: [EL-3395] move ApplicationAdvanceSettings in fsd structure
